infra/app/requirements.txt
pandas
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
